Add setup.py for backward compatibility
Python 3.6 in Fedora 37+ uses bundled wheels of setuptools and pip
because the recent versions packaged as python-*-wheel no longer support 3.6.

tox 3 creates an sdist and installs it into the py36 environment with
setuptools 40.6.2. setup.cfg-only support was added in 40.9.0.

This worked on Fedora 36 because Python 3.6 used the packaged wheels.